refactor : 번역 로직 수정&이미지 처리 로직 분리
